@@ -1,27 +1,42 @@
 package com.kishore.news
 
 import android.app.Application
-import com.kishore.news.common.dependency.koin.mySharedPreferenceModule
-import com.kishore.news.common.depndency.dagger.DaggerNewsComponent
-import com.kishore.news.common.depndency.dagger.NewsComponent
-import com.kishore.news.common.depndency.dagger.NewsSharedPreferencesModule
-import org.koin.android.ext.android.startKoin
+import androidx.hilt.work.HiltWorkerFactory
+import androidx.work.Configuration
+import com.google.firebase.FirebaseApp
+import com.google.firebase.crashlytics.FirebaseCrashlytics
+import com.kishore.news.util.NewsUtil
+import dagger.hilt.android.HiltAndroidApp
+import javax.inject.Inject
 
 
-class NewsApplication : Application() {
+@HiltAndroidApp
+class NewsApplication : Application(), Configuration.Provider {
 
-    lateinit var newsComponent: NewsComponent
+    @Inject
+    lateinit var workerFactory: HiltWorkerFactory
+
+    /*
+        Needed has we are using Hilt in worker manager.
+     */
+    override fun getWorkManagerConfiguration(): Configuration {
+        return Configuration.Builder()
+            .setWorkerFactory(workerFactory)
+            .build();
+    }
 
     override fun onCreate() {
         super.onCreate()
-        // Dagger
-        newsComponent = DaggerNewsComponent
-                .builder()
-                .newsSharedPreferencesModule(NewsSharedPreferencesModule(this.applicationContext)).build()
+        FirebaseCrashlytics.getInstance().setUserId(NewsUtil.getUniqueId(this))
+        FirebaseCrashlytics.getInstance().setCustomKey("userId",NewsUtil.getUniqueId(this))
 
-        //Koin
-        startKoin(this, listOf(mySharedPreferenceModule))
-
-
+       /* if (applicationContext.packageName.equals("com.kishore.news")) {
+            FirebaseCrashlytics.getInstance().setUserId(NewsUtil.getUniqueId(this))
+            FirebaseCrashlytics.getInstance().setCustomKey("userId",NewsUtil.getUniqueId(this))
+        } else {
+            val state: Boolean = false
+            FirebaseApp.getInstance().setDataCollectionDefaultEnabled(state)
+        }*/
     }
+
 }
